storages: dir field
