+++ table.go
@@ -1,11 +1,8 @@
 package simpdb
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
-	"os"
-	"path/filepath"
 )
 
 var (
@@ -24,86 +21,10 @@ type Entity interface {
 
 // Table is access point for storage of one entity type.
 type Table[E Entity] struct {
-	// filename with JSON map for the table
-	filename string
+	// files directory
+	dir string
 	// name of entity in the table
 	name string
-}
-
-func (t *Table[E]) ensureFileExists() error {
-	dir := filepath.Dir(t.filename)
-	if _, err := os.Stat(dir); err != nil {
-		if !os.IsNotExist(err) {
-			return fmt.Errorf("failed checking directory %s: %w", dir, err)
-		}
-
-		// TODO: mkdirall
-		if err := os.Mkdir(dir, 0755); err != nil {
-			return fmt.Errorf("failed creating directory %s: %w", dir, err)
-		}
-	}
-
-	if _, err := os.Stat(t.filename); err != nil {
-		if !os.IsNotExist(err) {
-			return fmt.Errorf("failed checking table file %s: %w", t.filename, err)
-		}
-
-		file, err := os.OpenFile(t.filename, os.O_CREATE|os.O_RDWR, 0644)
-		if err != nil {
-			return fmt.Errorf("failed creating table file %s: %w", t.filename, err)
-		}
-		defer file.Close()
-
-		if _, err := file.Write([]byte("{}")); err != nil {
-			return fmt.Errorf("failed filling table file %s with inital data: %w", t.filename, err)
-		}
-	}
-
-	return nil
-}
-
-// Read all records from table that satisfy predicate.
-func (t *Table[E]) Read(filter func(E) bool) (map[string]E, error) {
-	if err := t.ensureFileExists(); err != nil {
-		return nil, fmt.Errorf("read failed: %w", err)
-	}
-
-	bytes, err := os.ReadFile(t.filename)
-	if err != nil {
-		return nil, fmt.Errorf("read failed while reading file: %w", err)
-	}
-
-	var all map[string]E
-	if err := json.Unmarshal(bytes, &all); err != nil {
-		return nil, fmt.Errorf("read failed while decoding data: %w", err)
-	}
-
-	res := make(map[string]E, len(all))
-	for id, entity := range all {
-		if filter(entity) {
-			res[id] = entity
-		}
-	}
-
-	return res, nil
-}
-
-// Write fills table with entities.
-func (t *Table[E]) Write(entities map[string]E) error {
-	if err := t.ensureFileExists(); err != nil {
-		return fmt.Errorf("write failed: %w", err)
-	}
-
-	bytes, err := json.Marshal(entities)
-	if err != nil {
-		return fmt.Errorf("write failed while encoding json: %w", err)
-	}
-
-	if err := os.WriteFile(t.filename, bytes, 0644); err != nil {
-		return fmt.Errorf("write failed while writing file: %w", err)
-	}
-
-	return nil
 }
 
 // Update all records in table.
